(feat) Choose between Facebook and Meetup events for page display
The backend structure currently passes a callback into the
meetup api call. Once the call completes, the response is sent
and displayed on the page. Functionality is now included for
Facebook to replace Meetup. To have both sets of events display
on the page, the workflow will need to be adjusted.